requirements.txt
python-docx==1.1.2
pydantic==2.9.2
streamlit==1.38.0
python-slugify==8.0.4
tiktoken==0.7.0
